third_party/blink/web_tests/fast/js/instanceof-test-expected.txt

CONSOLE WARNING: XSLTProcessor and XSLT Processing Instructions have been deprecated by all browsers. These features will be removed from this browser soon. See https://chromestatus.com/feature/4709671889534976.
instanceof test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS (new Audio()) instanceof Audio is true
PASS (new Array()) instanceof Audio is false
PASS (new Image()) instanceof Image is true
PASS (new Array()) instanceof Image is false
PASS (new Option()) instanceof Option is true
PASS (new Array()) instanceof Option is false
PASS (new WebKitCSSMatrix()) instanceof WebKitCSSMatrix is true
PASS (new Array()) instanceof WebKitCSSMatrix is false
PASS (new Worker('instanceof-operator-dummy-worker.js')) instanceof Worker is true
PASS (new Array()) instanceof Worker is false
PASS (new XSLTProcessor()) instanceof XSLTProcessor is true
PASS (new Array()) instanceof XSLTProcessor is false
PASS successfullyParsed is true

TEST COMPLETE
